+++ src/vs/platform/dnd/browser/dnd.ts
@@ -358,3 +358,52 @@ export const Extensions = {
 Registry.add(Extensions.DragAndDropContribution, new DragAndDropContributionRegistry());
 
 //#endregion
+
+//#region DND Utilities
+
+/**
+ * A singleton to store transfer data during drag & drop operations that are only valid within the application.
+ */
+export class LocalSelectionTransfer<T> {
+
+	private static readonly INSTANCE = new LocalSelectionTransfer();
+
+	private data?: T[];
+	private proto?: T;
+
+	private constructor() {
+		// protect against external instantiation
+	}
+
+	static getInstance<T>(): LocalSelectionTransfer<T> {
+		return LocalSelectionTransfer.INSTANCE as LocalSelectionTransfer<T>;
+	}
+
+	hasData(proto: T): boolean {
+		return proto && proto === this.proto;
+	}
+
+	clearData(proto: T): void {
+		if (this.hasData(proto)) {
+			this.proto = undefined;
+			this.data = undefined;
+		}
+	}
+
+	getData(proto: T): T[] | undefined {
+		if (this.hasData(proto)) {
+			return this.data;
+		}
+
+		return undefined;
+	}
+
+	setData(data: T[], proto: T): void {
+		if (proto) {
+			this.data = data;
+			this.proto = proto;
+		}
+	}
+}
+
+//#endregion
